Renamed 'slice_config' to 'configuration'.
